put common lecture data into the data folder and fix imports
