+++ src/routes/Lectures.jsx
@@ -5,162 +5,12 @@ import {
   PlayCircleIcon,
   CalculatorIcon,
   DocumentDuplicateIcon
-} from '@heroicons/react/24/outline'
+} from '@heroicons/react/24/outline';
+
+import { lectureData as lectures } from '../assets/data/lectureData.js';
 
 export default function Lectures() {
 
-  const lectures = [
-    {
-      title: '01: Intro to Seismic Exam',
-      text: 'An overview of the Seismic Exam and its requirements. No structural content here, just background info on the exam itself and what to expect the day of.',
-      minutes: 30,
-      pages: 10,
-      problems: 0,
-      difficulty: 0,
-      img: '/images/lectures/01_Intro.png',
-      href: "01"
-    },
-    {
-      title: '02: Geology & Earthquakes',
-      text: 'General background information on tectonic activity and how eathquakes happen. A good starting point for understanding the hazards associated with seismic events and how it leads to building code requirements.',
-      minutes: 60,
-      pages: 10,
-      problems: 0,
-      difficulty: 0,
-      img: '/images/lectures/02_Geology.png',
-      href: "02"
-    },
-    {
-      title: '03: Code Provisions',
-      text: 'Introduction to applicable building codes in the United States including the IBC, and how it is used as a basis for the CBC. Necessary information on how ASCE 7 works and its relation to the CBC.',
-      minutes: 60,
-      pages: 10,
-      problems: 3,
-      difficulty: 0,
-      img: '/images/lectures/03_Code.png',
-      href: "03"
-    },
-    {
-      title: '04: Site Conditions',
-      text: 'Bridging the gap between general seismic maps and site specific conditions. Site factors based on soil characteristics.',
-      minutes: 60,
-      pages: 10,
-      problems: 3,
-      difficulty: 0,
-      img: '/images/lectures/04_Site.png',
-      href: "04"
-    },
-    {
-      title: '05: Risk Categories',
-      text: 'Explanation of seismic risk category for building structures per ASCE 7 and the implications on seismic performance. An essential concept tested on the exam, and a basic starting point of all seismic design.',
-      minutes: 60,
-      pages: 10,
-      problems: 5,
-      difficulty: 1,
-      img: '/images/lectures/05_Risk.png',
-      href: "05"
-    },
-    {
-      title: '06: Building Systems',
-      text: 'Diagrams and explanations of building systems including vertical resistand and lateral resisting systems. Basic calculations for lateral deflection, building periods, and degrees of freedom. This covers fundamentals taught in school that are assumed basic knowledge for using ASCE 7.',
-      minutes: 60,
-      pages: 10,
-      problems: 5,
-      difficulty: 1,
-      img: '/images/lectures/06_Building.png',
-      href: "06"
-    },
-    {
-      title: '07: Ductility',
-      text: 'How the concept of ductility relates to defining the magnitude of seismic forces with the Response Modification Factor, Deflection Amplification Factor, and Overstrength Factor. How to select building lateral systems and determine their factors using ASCE 7. Limitations with different lateral systems based on seismic design categories.',
-      minutes: 60,
-      pages: 21,
-      problems: 5,
-      difficulty: 1,
-      img: '/images/lectures/07_Ductility.png',
-      href: "07"
-    },
-    {
-      title: '08: Base Shear',
-      text: 'Determine seismic base shear for a building structure using the equivalent lateral force procedure in ASCE 7 with the response spectrum. How to categorize dead loads attributed to seismic mass, and approximate the fundamental period of a structure.',
-      minutes: 60,
-      pages: 23,
-      problems: 5,
-      difficulty: 2,
-      img: '/images/lectures/08_Base.png',
-      href: "08"
-    },
-    {
-      title: '09: Force Distribution',
-      text: 'Calculation process for determining Horizontal and Vertical Force Distribution along the height of a building structure. How to determine the seismic mass at individual levels in a building using tributary area methods. A relatively straighforward calculation procedure that will require practice to perform at speed. ',
-      minutes: 60,
-      pages: 31,
-      problems: 5,
-      difficulty: 2,
-      img: '/images/lectures/09_Force.png',
-      href: "09"
-    },
-    {
-      title: '10: Drift',
-      text: 'Drift calculation procedure and drift limits. Visual explanations of story drift versus story displacement, and how to use the correct seismic forces for drift checks. Additional information on buiding separation requirements and P-Delta effects.',
-      minutes: 60,
-      pages: 40,
-      problems: 5,
-      difficulty: 1,
-      img: '/images/lectures/10_Drift.png',
-      href: "10"
-    },
-    {
-      title: '11: Flexible Diaphragms',
-      text: 'Comprehensive introduction to the function and definition of building diaphragms. Calculation procedures for analyzing all relevant forces in flexible diaphragms with extensive examples. Collectors, chord forces, unit shear calculations. The second hardest thing to learn for the exam.',
-      minutes: 80,
-      pages: 71,
-      problems: 5,
-      difficulty: 3,
-      img: '/images/lectures/11_Flexible.png',
-      href: "11"
-    },
-    {
-      title: '12: Rigid Diaphragms',
-      text: 'Calculation procedure for rigid diaphragms. How the lateral resisting elements can change the force flow through a rigid diaphragm, and how to calculate relative rigidities of walls. Implementing accidental torsion per ASCE 7 requirements. The hardest thing to learn for the exam.',
-      minutes: 80,
-      pages: 60,
-      problems: 5,
-      difficulty: 3,
-      img: '/images/lectures/12_Rigid.png',
-      href: "12"
-    },
-    {
-      title: '13: Irregularities',
-      text: 'Classifying Vertical and Horizontal building irregularities per ASCE 7. Using irregularities to evaluate permitted analytical procedures. Determining seismic design force using redundancy or omega factors.',
-      minutes: 80,
-      pages: 32,
-      problems: 5,
-      difficulty: 1,
-      img: '/images/lectures/13_Irregularities.png',
-      href: "13"
-    },
-    {
-      title: '14: Non-structural Components',
-      text: 'Calculating forces on non-structural components. Determing overturning forces and brace forces for common components use basic static analysis. When to use Omega for concrete anchorage forces. Impacts of non-structural damage and costs during seismic events.',
-      minutes: 60,
-      pages: 43,
-      problems: 5,
-      difficulty: 2,
-      img: '/images/lectures/14_Nonstructural.png',
-      href: "14"
-    },
-    {
-      title: '15: Seismic Detailing',
-      text: 'Holistic review of load path for lateral elements. How to identify basic requirements for good force transfer across shearwalls and into foundations. Critical connections for diaphragms to walls.',
-      minutes: 40,
-      pages: 20,
-      problems: 5,
-      difficulty: 1,
-      img: '/images/lectures/15_Seismic.png',
-      href: "15"
-    },
-  ]
 
     return (
       <main>
